updated generate-3d.sh to: -- produce one *.mc file as the structure definition -- produce one *.xml file for adding texture -- produce one *.vti file for viewing -- produce a grain_stats.txt file for simple data analysis.
Additionally in the post processing the final grain structure will be
randomly numbered and the user will be able to threshold the grains
based on size. 

This script will also automatically checkout the voxelconvert code
from Latir and place it into the includes directory if it does not
already exist.